requirements.txt
langchain
langchain-google-genai
langchain-community
streamlit
python-dotenv
chromadb
pypdf
faiss-cpu
